feat(help): move help-window-toolbar inside help-window dir
