+++ src/platform/experimental/help/help-window/help-window-toolbar/help-window-toolbar.component.html
